Added Test-highOrderGrad.C
