+++ applications/test/highOrderGrad/Make/files
@@ -1,0 +1,3 @@
+Test-highOrderGrad.C
+
+EXE = $(FOAM_MODULE_APPBIN)/Test-highOrderGrad
