--- applications/test/kExactGrad/Make/files
@@ -1,3 +1,0 @@
-Test-kExactGrad.C
-
-EXE = $(FOAM_MODULE_APPBIN)/Test-kExactGrad
